+++ src/flags.h
@@ -64,6 +64,7 @@
 #define CBOR_MAP_AS_ARRAY     (1 << 6)
 
 /**
+ * Encode: Error if Traversable contains duplicate key.
  * Decode: Error if map key overwrites an existing key.
  */
 #define CBOR_MAP_NO_DUP_KEY   (1 << 7)